MyProject/Data_Collection/requirements.txt
numpy==1.21.2
pandas==1.3.2
psutil
paho-mqtt
psycopg2-binary